[PLACES] Add search criteria to nearby places
- Pending to implement Command to decouple requests to external api
- Pending to implement more tests
- Pending to test command coverage

+++ src/test/java/dev/corusoft/slurp/TestConstants.java
@@ -17,7 +17,7 @@ public class TestConstants {
     public final static String LOCALDATE_ISO_FORMAT = DateTimeFormat.ISO.DATE.name();
     public final static String DATETIME_FORMAT = "dd-MM-yyyy";
 
-    public final static double DEFAULT_SEARCH_PERIMETER_RADIUS = 200;
+    public final static int DEFAULT_SEARCH_PERIMETER_RADIUS = 200;
     public final static double MADRID_KILOMETRIC_POINT_0_LATITUDE = 40.4169473;
     public final static double MADRID_KILOMETRIC_POINT_0_LONGITUDE = -3.7035285;
 }
